chore(folder-structure): Rename files Rename files to use camelCase and remove plural definitions for some names
