+++ .env.example
@@ -2,3 +2,7 @@ NODE_ENV=
 PORT=
 GOOGLE_CLIENT_ID=
 GOOGLE_CLIENT_SECRET=
+
+POSTGRES_PASSWORD=
+POSTGRES_USER=
+POSTGRES_DB=
